feat: consolidate probability helpers
