+++ HISTORY.md
@@ -19,6 +19,9 @@
 -->
 
 # Version History
+- 0.2.171 - Consolidate probability and reliability helpers into
+          ``SafetyAnalysisService`` and remove ``probability_reliability``
+          module. Update AnalysisUtils to reference the unified service.
 - 0.2.170 - Wrap ``winfo_containing`` in ``try/except`` to guard ``KeyError``
           during drag target resolution and detach tabs safely when widgets
           vanish.  Add regression test simulating release over a destroyed
