+++ mainappsrc/version.py
@@ -18,6 +18,6 @@
 
 """Project version information."""
 
-VERSION = "0.2.170"
+VERSION = "0.2.171"
 
 __all__ = ["VERSION"]
